refactor: 폴더이름 변경 및 정리
