+++ vite-project/src/pages/FreeBoard-Page/FreeBoard.jsx
