+++ vite-project/src/pages/Items-Page/components/ProductBest/index.jsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from "react";
 import { getProduct } from "../../../../utils/api";
-import ProductCard from "../../../../common/productCard/ProductCard";
-import { Link } from "react-router";
+import ProductList from "../../../../common/Product-Card/ProductList";
+
 const ProductBest = () => {
   const [bestProducts, setBestProducts] = useState([]);
   useEffect(() => {
@@ -17,15 +17,7 @@ const ProductBest = () => {
   return (
     <div className="productBest-layout">
       <h1 className="product-category-description">베스트 상품</h1>
-      <ul className="product-best-list">
-        {bestProducts?.map((product) => (
-          <li key={product.id}>
-            <Link>
-              {<ProductCard product={product} category="best-card-img" />}
-            </Link>
-          </li>
-        ))}
-      </ul>
+      <ProductList allProducts={bestProducts} className="product-best-list" />
     </div>
   );
 };
